cleanup: move code where it belongs
