+++ fedimint-core/Cargo.toml
@@ -33,7 +33,6 @@ bitcoin-io = "0.1.2"
 bitcoin-units = "0.1.2"
 bitvec = "1.0.1"
 bls12_381 = { workspace = true }
-bs58 = "0.5.1"
 erased-serde = { workspace = true }
 fedimint-derive = { workspace = true }
 fedimint-logging = { workspace = true }
