+++ fedimint-core/src/admin_client.rs
@@ -1,14 +1,11 @@
 use std::collections::BTreeMap;
 use std::fmt::Debug;
 
-use anyhow::ensure;
 use fedimint_core::util::SafeUrl;
 use serde::{Deserialize, Serialize};
 
 use crate::config::ServerModuleConfigGenParamsRegistry;
 use crate::encoding::{Decodable, Encodable};
-use crate::module::registry::ModuleDecoderRegistry;
-use crate::PeerId;
 
 /// The state of the server returned via APIs
 #[derive(Debug, Clone, Default, Serialize, Deserialize, Eq, PartialEq, Encodable, Decodable)]
@@ -75,53 +72,6 @@ pub struct SetLocalParamsRequest {
     pub name: String,
     /// Federation name set by the leader
     pub federation_name: Option<String>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Encodable, Decodable)]
-/// Connection information sent between peers in order to start config gen
-pub struct PeerConnectionInfo {
-    /// TLS cert is necessary for P2P auth during DKG and consensus
-    pub cert: Vec<u8>,
-    /// P2P is the network for running DKG and consensus
-    pub p2p_url: SafeUrl,
-    /// API for secure websocket requests
-    pub api_url: SafeUrl,
-    /// Name of the peer, used in TLS auth
-    pub name: String,
-    /// Federation name set by the leader
-    pub federation_name: Option<String>,
-}
-
-impl PeerConnectionInfo {
-    pub fn encode_base58(&self) -> String {
-        format!(
-            "fedimint{}",
-            bs58::encode(&self.consensus_encode_to_vec()).into_string()
-        )
-    }
-
-    pub fn decode_base58(s: &str) -> anyhow::Result<Self> {
-        ensure!(s.starts_with("fedimint"), "Invalid Prefix");
-
-        let params = Self::consensus_decode_whole(
-            &bs58::decode(&s[8..]).into_vec()?,
-            &ModuleDecoderRegistry::default(),
-        )?;
-
-        Ok(params)
-    }
-}
-
-/// The config gen params that need to be in consensus, sent by the config gen
-/// leader to all the other guardians
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct ConfigGenParamsConsensus {
-    /// Endpoints of all servers
-    pub peers: BTreeMap<PeerId, PeerConnectionInfo>,
-    /// Guardian-defined key-value pairs that will be passed to the client
-    pub meta: BTreeMap<String, String>,
-    /// Module init params (also contains local params from us)
-    pub modules: ServerModuleConfigGenParamsRegistry,
 }
 
 /// Config gen params that can be configured from the UI
